Spring security & Authentication

+++ src/main/java/com/example/journalApp/service/JournalEntryService.java
@@ -7,6 +7,7 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -22,11 +23,13 @@ public class JournalEntryService {
     @Autowired
     public UserService userService;
 
+    @Transactional
     public void saveEntryByUsername(JournalEntry journalEntry, String username){
         User user = userService.findByUsername(username);
         journalEntry.setDate(LocalDateTime.now());
         JournalEntry saved = journalEntryRepository.save(journalEntry);
         user.getJournalEntries().add(saved);
+//        user.setUsername(null);
         userService.saveEntry(user);
     }
 
@@ -42,10 +45,14 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String username){
+    public Boolean deleteById(ObjectId id, String username){
+        boolean removed = false;
         User user = userService.findByUsername(username);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if(removed){
+            userService.saveEntry(user);
+            journalEntryRepository.deleteById(id);
+        }
+        return removed;
     }
 }
